refactor: 폴더구조 변경 및 Subtitle 분리
- Title 파일을 register 에서 common 폴더로 이동
- 로그인, 회원가입의 Title 을 Subtitle, Title 로 분리
