+++ src/components/common/Title.jsx
@@ -8,11 +8,10 @@ const Title = ({ children, fontSize, margin }) => {
   );
 };
 
-const TitleStyle = styled.p`
+const TitleStyle = styled.h3`
   color: #216d32;
   font-size: ${(props) => props.fontSize};
   margin: ${(props) => props.margin};
-  /* margin-bottom: 4rem; */
 `;
 
 export default Title;
